bug 1148476, verbatim-codes-check with locale teams and BeautifulSoup, rs=foopy
Rewrite the command to not depend on pyquery, also remove the dependency.
The code also didn't know about the locale-team mappings, that's added now.
Also moves from using print to self.stdout, like django commands should.

I also took the liberty to rewrite the logic until I understood what it
does again. It now generates a proposed VERBATIM_CONVERSIONS mapping as
it goes, and compares that to the reference.