bug 1150466, transactions and other ORM fixes for buildbot interactions
Only use transactions for DB constructs that are complex, notably
changes (with files and properties) and sourcestamps (with changes, sic).

Everything else is fine with just autocommit, which we now have with
Django 1.6.

This is part of the bigger update work in bug 1141023.

Also, this patch is a bit more precise on when and if to save objects
after modifying manytomany relations. 'Cause then you don't have to.