Create initial data for life in the initial migration, more natural ordering of apps
Loading the initial data for life is brittle for migration and app order, at
least starting with Django 1.6.
Thus I'm moving the initial data into the initial migration, and remove
the initial_data.json. That's deprecated in Django 1.8 anyway, so we're getting
ahead of ourselves here.

Also reordering the elmo apps a bit so that they're more in order of dependencies.