bug 1169346, show fallbacks in about-milestone and locale_changes.
Prior behavior was to show fallbacks as additions, which confused the
difference between old and new.
For the locale_changes, the trick is to reduce the milestone_signoffs query
to mappings where both the signoff and the milestone are for the current
appversion.
For about-milestone, we need to get the fallbacks listed, and exclude them
from the added data. For a better overview of what's new and what's old,
I'm showing them in an extra column.