bug 1188833, rework fallback logic in shipping.api
There's a bunch of fixes in this patch.
Firstly, if we got an appversion covered, don't evaluate it again.
Then, once we're falling back, make sure we're only trying locales
that actually had an ACCEPTED signoff in our fallback chain.
That way, our iteration actually comes to a rather early and good
end.