Update script to fetch, prune obsolete pyc, merge
When refactoring code, we often see left-over *.pyc files, which
keep the modules around, and the dirs aren't clobbered.

Best way I could find to get around that is to fetch, check
which files get deleted between HEAD and FETCH_HEAD, remove
the corresponding pyc files, and then merge.

I also refactored the script to include the source update
scripts in the parent script, and to just have the install and
bootstrap scripts in the external python module, which is then
imported from the code that we update to.