bug 1195770, TEMPLATES setting instead of TEMPLATE_, deprecations.
This patch is bigger than it should be because the tests to
ensure that session_csrf is what we're using were kinda wonky.

In particular, django hard-codes that you have
django.template.context_processors.csrf on. (Used to be django.core..)
session_csrf works because it comes *after* it, and overrides the
csrf token with its own version. TIL.