bug 1330707, update django in automation needs elmo fixes.
There are two culprits left in the bb2mbdb bridge code:

We reload() django modules, which breaks django's assumptions
on python class objects staying the same throughout the lifetime.

And we did django.setup() somewhere in the middle, which should
really only happen once at the main entry points. In this case,
in the buildmaster's shared.cfg.