bug 1290906, resolve default or tip to the last known default changeset in the db, r=rail
Resolving 'default' is slow, and as soon as we use unified repos on elmo, also
wrong. Let's pick the last changeset of the default branch from the db, and use
an explicit version when talking to hg.
I'm also aliasing this to tip, which is not strictly true always, but it
shouldn't be used to begin with.
Better fix would be to just not have this code path at all, which is
bug 1280730.