bug 1288656, show a checkbox to accept sign-off, f=flod
Show a checkbox to accept a Sign-Off while signing off.
Only show it to folks with the right permission, and only
check the flag for the same folks in the python code, too.

I default to "on", just because I think it's the most common action.
For rejecting a revision, you still have to go through the double
step of signing off and then rejecting. Might be OK, as you
want a message in that case, too.