bug 1429410, follow-up, run ChangeMaster.nextNumber reset always
We actually need to sync the nextNumber on each startup.

There are two reasons for loosing data:
- mbdb looses obsolete Changes, freeing up entries in the db.
- buildbot looses the last changenumber due to unclean shutdown.

The first case, mbdb loosing data, we just ignore. IDs are cheap.
The latter case, buildbot would try to use the numbers in our
db again, causing all kind of trouble.

In that case, we set the nextNumber. We also clear out the existing
changes that ChangeMaster synced back, as some parts of buildbot
code assume that the changes are stored consecutive.
That's harmless, and our changeHorizon is 1 anyway by now, so
we don't have much data anymore in ChangeMaster.