bug 1484189, make progress graphs independent of the db
We used to put the background positions into the db, but for
multiple web heads, that's a bad idea.
Instead generate a css file along with the png image, and use
that.
If we move from file-based assets to generated views, this
will be a good head start.

progress.css isn't available on test time, whitelist that when
we test our includes.