bug 1520769, clean up ContentType and permissions in db through migration.
We've taken a couple of migrations that removed apps and models that
didn't remove the content types or the permissions for them. Remove
them now.

The migration depends on that latest migration of all elmo apps with
data, so it's OK to access them virtually in the migration via get_model.