Convert progress spark from static files into a dynamic CSS.
This is OK now that the progress view is faster. To ensure that
the CSS and the image are aligned, we're putting both into a single
resource, by using a data uri in the CSS.

Also, load this stylesheet lazily to not block rendering, that's not required.

The layout of the thumbs is extracted into a different stylesheet.
That's still rendered as it's settings-dependent, but loaded blockingly.

As we now have a view for progress.css, remove the old static from
the WHITELIST in the includes test.