Fix resource-check test.
This test tries to ensure that the included resources are actually somewhere.
It did so rather poorly, for example fat-fingering foo.css as foo.csss
didn't make this test fail.
We're including staticfiles unconditionally these days, so use it in our
tests, never actually call .get() to speed up tests, and use
urlparse() which is the first thing that .get() does.
This variety of the test catches a ton more resources, in particular js
with query strings.