bug 1596489, switch on cert validation with certifi.where.
This replaces the options used to create the Elasticsearch
instance from ES_COMPARE_HOST to ES_KWARGS, which is derived
from ES_COMPARE_HOST in elmo.settings. That way, we can
ensure consistent parameters passed to ES, regardless whether
it's the view or the heartbeat or whatnot.

Side note, we must not set `use_ssl`, as that's per host.
If we pass it in to ElasticSearch, non-ssl localhost/docker-compose
breaks.